agent/requirements.txt
psutil==6.0.0
Pillow==10.4.0
pytesseract==0.3.13
